move some identity based functions to identity.php
